added exceptions to all delete methods and changed deleteCountryLanguage to take a string language input

+++ src/main/java/com/sparta/jarjarbinks/worldproject/model/services/WorldService.java
@@ -1,5 +1,7 @@
 package com.sparta.jarjarbinks.worldproject.model.services;
 
+import com.sparta.jarjarbinks.worldproject.exceptions.InvalidArgumentFormatException;
+import com.sparta.jarjarbinks.worldproject.exceptions.NotFoundException;
 import com.sparta.jarjarbinks.worldproject.model.entities.CityDTO;
 import com.sparta.jarjarbinks.worldproject.model.entities.CountryDTO;
 import com.sparta.jarjarbinks.worldproject.model.entities.CountrylanguageDTO;
@@ -193,20 +195,48 @@ public class WorldService {
         return  percentage.multiply(BigDecimal.valueOf(countryDTO.getPopulation())).intValue();
     }
 
-    public void deleteCity(Integer id){
-        cityRepository.deleteById(id);
+    public void deleteCity(Integer id) throws NotFoundException, InvalidArgumentFormatException {
+        if(id == null){
+            throw new InvalidArgumentFormatException("Invalid input into deleteCity method");
+        }
+        if(cityRepository.findById(id).isEmpty()){
+            throw new NotFoundException("This city could not be found.");
+        }else {
+            cityRepository.deleteById(id);
+        }
     }
-    public void deleteCountry(Integer id){
-        Optional<CountryDTO> thisCountry = countryRepository.findById(id.toString());
-        List<CityDTO> thisCity = cityRepository.findAllByCountryCode(thisCountry);
+    public void deleteCountry(Integer countryId) throws InvalidArgumentFormatException, NotFoundException {
 
-        if(thisCountry.isPresent()){
+        if(countryId == null){
+            throw new InvalidArgumentFormatException("Invalid input into deleteCountry method");
+        }
+
+        if(countryRepository.findById(countryId.toString()).isEmpty()){
+            throw new NotFoundException("This Country could not be found.");
+        } else {
+            Optional<CountryDTO> thisCountry = countryRepository.findById(countryId.toString());
+            List<CityDTO> thisCity = cityRepository.findAllByCountryCode(thisCountry);
             countryRepository.delete(thisCountry);
             cityRepository.deleteAllByCountryCode(thisCountry);
         }
+
     }
 
-    public void deleteCountryLanguage(CountrylanguageIdDTO countrylanguageIdDTO){
-        countrylanguageRepository.deleteById(countrylanguageIdDTO);
+    public void deleteCountryLanguage(String language) throws InvalidArgumentFormatException, NotFoundException {
+        if(language == null){
+            throw new InvalidArgumentFormatException("Invalid language was entered make sure the language is Capitalized");
+        }
+        int count = 0;
+
+        List<CountrylanguageDTO> languages = countrylanguageRepository.findAll();
+        for(CountrylanguageDTO country: languages){
+            if(country.getId().getLanguage().equals(language)){
+                countrylanguageRepository.delete(country);
+                count++;
+            }
+        }
+        if (count == 0){
+            throw new NotFoundException("The Language " + language + " could not be found.");
+        }
     }
 }
